Complete 04-chaining-selectors
Chaining selectors is when we list them without any separation. This
will select any elements that has both selectors. This syntax will
work for chaining any combination of selectors except for chaining
more than one "type" selectors. Ex (div and p)

When adjusting width or height of an image, if you want to adjust
the size without losing its proportions, use "auto" for the unedited
side.

It is best to include both width and height of an image as it helps
the website to figure out how to display while the image loads.